requirements.txt
aiogram==2.22.1
aiohttp==3.8.3
aiosignal==1.2.0
altair==4.2.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
backports.zoneinfo==0.2.1
bcrypt==4.0.0
blinker==1.5
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
cmake==3.24.1.1
commonmark==0.9.1
contourpy==1.0.5
cycler==0.11.0
decorator==5.1.1
deta==1.1.0
entrypoints==0.4
et-xmlfile==1.1.0
extra-streamlit-components==0.1.55
fonttools==4.37.3
frozenlist==1.3.1
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.10.1
google-api-python-client==2.62.0
google-auth==2.11.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.4
importlib-metadata==4.12.0
importlib-resources==5.9.0
Jinja2==3.1.2
jsonschema==4.16.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
multidict==6.0.2
numpy==1.23.3
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.1
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.8.0b3
Pygments==2.13.0
PyJWT==2.5.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
rich==12.5.1
rsa==4.9
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.13.0
streamlit-authenticator==0.1.5
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
validators==0.20.0
yarl==1.8.1
zipp==3.8.1
